@@ -12,38 +12,39 @@
   },
   "private": true,
   "dependencies": {
+    "@angular-devkit/build-angular": "~0.5.0",
     "@angular/animations": "^5.2.0",
+    "@angular/cli": "^6.0.0-rc.2",
     "@angular/common": "^5.2.0",
     "@angular/compiler": "^5.2.0",
+    "@angular/compiler-cli": "^5.2.0",
     "@angular/core": "^5.2.0",
     "@angular/forms": "^5.2.0",
     "@angular/http": "^5.2.0",
+    "@angular/language-service": "^5.2.0",
     "@angular/platform-browser": "^5.2.0",
     "@angular/platform-browser-dynamic": "^5.2.0",
     "@angular/router": "^5.2.0",
-    "angular-in-memory-web-api": "^0.5.3",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
-  },
-  "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@ngtools/webpack": "6.0.0-rc.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "^9.6.2",
+    "angular-in-memory-web-api": "0.5.3",
     "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "core-js": "^2.5.4",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "protractor": "^5.3.1",
+    "rxjs": "^6.0.0-tactical-rc.1",
+    "rxjs-compat": "^6.0.0-tactical-rc.1",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.8.1",
+    "zone.js": "^0.8.25"
   }
 }
