@@ -4,7 +4,7 @@
 import { Component } from '@angular/core';
 import { FormControl } from '@angular/forms';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import {
   concatMap, catchError,
