@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Movie } from 'app/samples/sw-interfaces';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 // Add each RxJS operator to Observable
 import 'rxjs/add/operator/catch'; // now `catchError`
