@@ -4,7 +4,7 @@ import { HttpClient } from '@angular/common/http';
 import { RootMovies, Movie } from '../sw-interfaces';
 import { SwUrlService } from 'app/samples/sw-url.service';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 // As of v5.5: import symbols from operators
 import { catchError, map } from 'rxjs/operators';
