@@ -1,16 +1,21 @@
-  // tslint:disable:member-ordering
-import { Injectable } from '@angular/core';
+// tslint:disable:member-ordering
 import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
 
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
-import { of } from 'rxjs/observable/of';
-
+import { Observable, ReplaySubject, of } from 'rxjs';
 import {
-  catchError, concat, distinctUntilChanged,
-  map, merge, mergeMap,
-  scan, shareReplay, startWith, switchMap, tap } from 'rxjs/operators';
+  catchError,
+  concat,
+  distinctUntilChanged,
+  map,
+  merge,
+  mergeMap,
+  scan,
+  shareReplay,
+  startWith,
+  switchMap,
+  tap
+} from 'rxjs/operators';
 
 import { People, RootPeople } from 'app/samples/sw-interfaces';
 
@@ -18,7 +23,6 @@ const log = (desc: string) => (...args) => console.log(desc, ...args);
 
 @Injectable()
 export class SwPeopleService {
-
   private addOne$ = new ReplaySubject<People[]>(Number.POSITIVE_INFINITY);
 
   private additions$ = this.addOne$.pipe(
@@ -27,13 +31,13 @@ export class SwPeopleService {
   );
 
   private load = (url: string): Observable<RootPeople> =>
-    this.http.get<RootPeople>(url).pipe(
-      catchError(_ => of<RootPeople>(null)),
-      tap(() => console.log('fetched from back-end: ' + url)),
-      switchMap(r => r.next ?
-            of(r).pipe(merge(this.load(r.next))) :
-            of(r))
-    );
+    this.http
+      .get<RootPeople>(url)
+      .pipe(
+        catchError(_ => of<RootPeople>(null)),
+        tap(() => console.log('fetched from back-end: ' + url)),
+        switchMap(r => (r.next ? of(r).pipe(merge(this.load(r.next))) : of(r)))
+      );
 
   private _people$ = of(`https://swapi.co/api/people/`).pipe(
     mergeMap(this.load),
@@ -43,24 +47,18 @@ export class SwPeopleService {
   );
 
   people$ = this._people$.pipe(
-      mergeMap(list => this.additions$.pipe(
-      map(additions => list.concat(additions)))
+    mergeMap(list =>
+      this.additions$.pipe(map(additions => list.concat(additions)))
     ),
     map(list => list.sort((x, y) => (x.name < y.name ? -1 : 1)))
   );
 
   loading$ = of(true).pipe(
-    merge(this._people$.pipe(
-      map(_ => true),
-      concat(of(false)))
-    ),
+    merge(this._people$.pipe(map(_ => true), concat(of(false)))),
     distinctUntilChanged()
   );
 
-  count$ = this.people$.pipe(
-    map((list: any[]) => list.length),
-    startWith(0)
-  );
+  count$ = this.people$.pipe(map((list: any[]) => list.length), startWith(0));
 
   find(partial) {
     return this.people$.pipe(
@@ -79,13 +77,13 @@ export class SwPeopleService {
     // generate an id! just for demo purposes.
     newCharacter.url = new Date().getTime().toString(36);
     /**
-         * Make sure you do the actual save to DB in here.
-         * as this function now only updates the local
-         */
+     * Make sure you do the actual save to DB in here.
+     * as this function now only updates the local
+     */
     return this.addOne$.next([newCharacter]);
   }
 
   save(update: People) {}
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 }
