@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 import { WikipediaService, WikiResult } from 'app/samples/wikipedia.service';
 import { FormControl } from '@angular/forms';
 
-import { of } from 'rxjs/observable/of';
+import { of } from 'rxjs';
 import {
   catchError,
   debounceTime,
